// src/openapi/model/api-project.ts
/* tslint:disable */
/* eslint-disable */
/**
 * esint_server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import {ApiGenre} from './api-genre';

/**
 * 
 * @export
 * @interface ApiProject
 */
export interface ApiProject {
    /**
     *
     * @type {number}
     * @memberof ApiProject
     */
    id?: number;
    /**
     *
     * @type {string}
     * @memberof ApiProject
     */
    createdAt?: string;
    /**
     *
     * @type {string}
     * @memberof ApiProject
     */
    updatedAt?: string;
    /**
     *
     * @type {string}
     * @memberof ApiProject
     */
    name: string;
    /**
     *
     * @type {number}
     * @memberof ApiProject
     */
    ordererId: number;
    /**
     *
     * @type {number}
     * @memberof ApiProject
     */
    areaId: number;
    /**
     *
     * @type {string}
     * @memberof ApiProject
     */
    areaName?: string;
    /**
     *
     * @type {number}
     * @memberof ApiProject
     */
    status?: number;
    /**
     *
     * @type {number}
     * @memberof ApiProject
     */
    feePercent?: number;
    /**
     * 
     * @type {number}
     * @memberof ApiProject
     */
    stock: number;
    /**
     * 
     * @type {string}
     * @memberof ApiProject
     */
    workStartDate: string;
    /**
     * 
     * @type {string}
     * @memberof ApiProject
     */
    workEndDate: string;
    /**
     * 
     * @type {string}
     * @memberof ApiProject
     */
    restraintStartDate: string;
    /**
     * 
     * @type {string}
     * @memberof ApiProject
     */
    restraintEndDate: string;
    /**
     * 
     * @type {string}
     * @memberof ApiProject
     */
    postalCode1: string;
    /**
     * 
     * @type {string}
     * @memberof ApiProject
     */
    postalCode2: string;
    /**
     * 
     * @type {string}
     * @memberof ApiProject
     */
    address1?: string;
    /**
     * 
     * @type {string}
     * @memberof ApiProject
     */
    address2?: string;
    /**
     * 
     * @type {number}
     * @memberof ApiProject
     */
    minPrice: number;
    /**
     * 
     * @type {number}
     * @memberof ApiProject
     */
    maxPrice: number;
    /**
     * 
     * @type {string}
     * @memberof ApiProject
     */
    description?: string;
    /**
     *
     * @type {string}
     * @memberof ApiProject
     */
    recruitingStartDate: string;
    /**
     *
     * @type {string}
     * @memberof ApiProject
     */
    recruitingEndDate: string;
    /**
     *
     * @type {number}
     * @memberof ApiProject
     */
    minimumExperienceYears?: number;
    /**
     *
     * @type {Array<ApiGenre>}
     * @memberof ApiProject
     */
    genres?: Array<ApiGenre>;
}